add day3 to qdc_2024_2 folder
