+++ qc_grader/challenges/qdc_2024/__init__.py
@@ -20,14 +20,3 @@ from .day2 import (
     grade_day2a_ex5,
     grade_day2b_ex1,
 )
-
-from .day3 import (
-    grade_day3a_ex1,
-    grade_day3a_ex2,
-    grade_day3a_ex3,
-    grade_day3a_ex4,
-    grade_day3a_ex5,
-    grade_day3b_ex1,
-    submit_day3a_ex5,
-    submit_day3b_ex1
-)
